phase-5/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "AhmedSaeed4", email = "ahmedsaeed20026@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "cryptography>=44.0.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "mcp>=1.25.0",
    "openai>=2.15.0",
    "openai-agents>=0.6.5",
    "openai-chatkit>=1.5.3",
    "python-dateutil>=2.8.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
